requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-environ==0.11.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
